requirements.txt
altair==5.5.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fire==0.7.0
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.21.1
numpy==2.2.1
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pdf2docx==0.5.8
pillow==11.1.0
protobuf==5.29.2
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.25.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
streamlit==1.41.1
streamlit-draggable-list==0.0.1
streamlit-pdf-viewer==0.0.20
tenacity==9.0.0
termcolor==2.5.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
streamlit_image_coordinates
